+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # tcping2
+## [1.1.1 - 2024-05-03]
+- move to golang 1.22
+- update dependencies
+- move Dockerfile to folder docker/image
 
 ## [1.1.0 - 2024-04-28]
 - added echo Server and Client and related tests